Rename package from str to nstring
