+++ database/flow_network.go
@@ -7,8 +7,8 @@ import (
 	"github.com/NubeIO/flow-framework/src/client"
 	"github.com/NubeIO/flow-framework/utils/boolean"
 	"github.com/NubeIO/flow-framework/utils/integer"
+	"github.com/NubeIO/flow-framework/utils/nstring"
 	"github.com/NubeIO/flow-framework/utils/nuuid"
-	"github.com/NubeIO/flow-framework/utils/str"
 	"github.com/NubeIO/nubeio-rubix-lib-models-go/pkg/v1/model"
 	log "github.com/sirupsen/logrus"
 	"gorm.io/gorm"
@@ -132,7 +132,7 @@ func (d *GormDatabase) RefreshFlowNetworksConnections() (*bool, error) {
 		fnModel := model.FlowNetworkClone{}
 		if err != nil {
 			fnModel.IsError = boolean.NewTrue()
-			fnModel.ErrorMsg = str.NewStringAddress(err.Error())
+			fnModel.ErrorMsg = nstring.NewStringAddress(err.Error())
 			fnModel.FlowToken = fn.FlowToken
 		} else {
 			fnModel.IsError = boolean.NewFalse()
@@ -191,7 +191,7 @@ func (d *GormDatabase) editFlowNetworkBody(body *model.FlowNetwork) (bool, *clie
 			body.IsRemote = boolean.NewFalse()
 			if !isMasterSlave {
 				body.FlowHTTPS = boolean.NewFalse()
-				body.FlowIP = str.NewStringAddress("0.0.0.0")
+				body.FlowIP = nstring.NewStringAddress("0.0.0.0")
 				body.IsRemote = boolean.NewFalse()
 			}
 		}
@@ -217,11 +217,11 @@ func (d *GormDatabase) afterCreateUpdateFlowNetwork(body *model.FlowNetwork, isM
 			return nil, err
 		}
 		bodyToSync.FlowHTTPS = localStorageFlowNetwork.FlowHTTPS
-		bodyToSync.FlowIP = str.NewStringAddress(localStorageFlowNetwork.FlowIP)
+		bodyToSync.FlowIP = nstring.NewStringAddress(localStorageFlowNetwork.FlowIP)
 		bodyToSync.FlowPort = integer.New(localStorageFlowNetwork.FlowPort)
-		bodyToSync.FlowUsername = str.NewStringAddress(localStorageFlowNetwork.FlowUsername)
-		bodyToSync.FlowPassword = str.NewStringAddress(localStorageFlowNetwork.FlowPassword)
-		bodyToSync.FlowToken = str.NewStringAddress(localStorageFlowNetwork.FlowToken)
+		bodyToSync.FlowUsername = nstring.NewStringAddress(localStorageFlowNetwork.FlowUsername)
+		bodyToSync.FlowPassword = nstring.NewStringAddress(localStorageFlowNetwork.FlowPassword)
+		bodyToSync.FlowToken = nstring.NewStringAddress(localStorageFlowNetwork.FlowToken)
 	}
 	err := d.syncAndEditFlowNetwork(cli, body, &bodyToSync)
 	if err != nil {
