+++ database/producer.go
@@ -5,8 +5,8 @@ import (
 	"github.com/NubeIO/flow-framework/api"
 	"github.com/NubeIO/flow-framework/src/client"
 	"github.com/NubeIO/flow-framework/utils/boolean"
+	"github.com/NubeIO/flow-framework/utils/nstring"
 	"github.com/NubeIO/flow-framework/utils/nuuid"
-	"github.com/NubeIO/flow-framework/utils/str"
 	"github.com/NubeIO/nubeio-rubix-lib-models-go/pkg/v1/model"
 	log "github.com/sirupsen/logrus"
 	"time"
@@ -202,7 +202,7 @@ func (d *GormDatabase) producerScheduleWrite(uuid string, scheduleData *model.Sc
 }
 
 func (d *GormDatabase) TriggerCOVToWriterClone(producer *model.Producer, body *model.SyncCOV) error {
-	wcs, err := d.GetWriterClones(api.Args{ProducerUUID: str.NewStringAddress(producer.UUID)})
+	wcs, err := d.GetWriterClones(api.Args{ProducerUUID: nstring.NewStringAddress(producer.UUID)})
 	if err != nil {
 		return errors.New("error on getting writer clones from producer_uuid")
 	}
