+++ database/sync_flow_network.go
@@ -5,8 +5,8 @@ import (
 	"errors"
 	"github.com/NubeIO/flow-framework/src/client"
 	"github.com/NubeIO/flow-framework/utils/boolean"
+	"github.com/NubeIO/flow-framework/utils/nstring"
 	"github.com/NubeIO/flow-framework/utils/nuuid"
-	"github.com/NubeIO/flow-framework/utils/str"
 	"github.com/NubeIO/nubeio-rubix-lib-models-go/pkg/v1/model"
 )
 
@@ -38,7 +38,7 @@ func (d *GormDatabase) SyncFlowNetwork(body *model.FlowNetwork) (*model.FlowNetw
 	fnc.SyncUUID, _ = nuuid.MakeUUID()
 	if !boolean.IsTrue(fnc.IsRemote) {
 		fnc.FlowHTTPS = boolean.NewFalse()
-		fnc.FlowIP = str.NewStringAddress("0.0.0.0")
+		fnc.FlowIP = nstring.NewStringAddress("0.0.0.0")
 	}
 	deviceInfo, err := d.GetDeviceInfo()
 	if err != nil {
