+++ database/writer.go
@@ -5,8 +5,8 @@ import (
 	"errors"
 	"github.com/NubeIO/flow-framework/api"
 	"github.com/NubeIO/flow-framework/src/client"
+	"github.com/NubeIO/flow-framework/utils/nstring"
 	"github.com/NubeIO/flow-framework/utils/nuuid"
-	"github.com/NubeIO/flow-framework/utils/str"
 	"github.com/NubeIO/nubeio-rubix-lib-models-go/pkg/v1/model"
 	log "github.com/sirupsen/logrus"
 	"gorm.io/datatypes"
@@ -124,7 +124,7 @@ func (d *GormDatabase) WriterAction(uuid string, body *model.WriterBody) *model.
 	output.Action = body.Action
 	if err != nil {
 		output.IsError = true
-		output.Message = str.NewStringAddress(err.Error())
+		output.Message = nstring.NewStringAddress(err.Error())
 		return output
 	}
 	consumer, _ := d.GetConsumer(writer.ConsumerUUID, api.Args{})
@@ -135,12 +135,12 @@ func (d *GormDatabase) WriterAction(uuid string, body *model.WriterBody) *model.
 		err = cli.SyncWriterReadAction(uuid)
 		if err != nil {
 			output.IsError = true
-			output.Message = str.NewStringAddress(err.Error())
+			output.Message = nstring.NewStringAddress(err.Error())
 		}
 		dataStore, presentValue, err := d.getDataStoreAndPresentValues(writer)
 		if err != nil {
 			output.IsError = true
-			output.Message = str.NewStringAddress(err.Error())
+			output.Message = nstring.NewStringAddress(err.Error())
 		}
 		output.DataStore = dataStore
 		output.PresentValue = presentValue
@@ -149,7 +149,7 @@ func (d *GormDatabase) WriterAction(uuid string, body *model.WriterBody) *model.
 		dataStore, presentValue, err := d.getDataStoreAndPresentValues(writer)
 		if err != nil {
 			output.IsError = true
-			output.Message = str.NewStringAddress(err.Error())
+			output.Message = nstring.NewStringAddress(err.Error())
 		}
 		output.DataStore = dataStore
 		output.PresentValue = presentValue
@@ -158,7 +158,7 @@ func (d *GormDatabase) WriterAction(uuid string, body *model.WriterBody) *model.
 		bytes, err := d.validateWriterWriteBody(writer.WriterThingClass, body)
 		if err != nil {
 			output.IsError = true
-			output.Message = str.NewStringAddress(err.Error())
+			output.Message = nstring.NewStringAddress(err.Error())
 			return output
 		}
 		writer.DataStore = bytes
@@ -170,13 +170,13 @@ func (d *GormDatabase) WriterAction(uuid string, body *model.WriterBody) *model.
 		err = cli.SyncWriterWriteAction(uuid, &syncWriterAction)
 		if err != nil {
 			output.IsError = true
-			output.Message = str.NewStringAddress(err.Error())
+			output.Message = nstring.NewStringAddress(err.Error())
 			return output
 		}
 		dataStore, presentValue, err := d.getDataStoreAndPresentValues(writer)
 		if err != nil {
 			output.IsError = true
-			output.Message = str.NewStringAddress(err.Error())
+			output.Message = nstring.NewStringAddress(err.Error())
 		}
 		output.DataStore = dataStore
 		output.PresentValue = presentValue
