+++ plugin/nube/protocals/lora/utils.go
@@ -3,7 +3,7 @@ package main
 import (
 	"github.com/NubeIO/flow-framework/utils/array"
 	"github.com/NubeIO/flow-framework/utils/integer"
-	"github.com/NubeIO/flow-framework/utils/str"
+	"github.com/NubeIO/flow-framework/utils/nstring"
 	"reflect"
 	"strings"
 
@@ -29,7 +29,7 @@ func (inst *Instance) wizardSerial(body wizard) (string, error) {
 	net.Name = model.TransProtocol.Lora
 	net.TransportType = model.TransType.Serial
 	net.PluginPath = model.TransProtocol.Lora
-	net.SerialPort = str.NewStringAddress(sp)
+	net.SerialPort = nstring.NewStringAddress(sp)
 	net.SerialBaudRate = integer.NewUint(38400)
 
 	var dev model.Device
