+++ utils/nstring/string.go
@@ -1,4 +1,4 @@
-package str
+package nstring
 
 import (
 	"errors"
